Drive-by: Add missing includes to param-names-ast.c++.
I noticed this list was out-of-date compared to rtti.c++.

Note I couldn't add modules.h nor pyodide.h as they don't seem to be declare dependencies. Unclear if that's important.

This doesn't seem to have had any effect on the generated types, so I'm not sure whether it matters, but the comment seemed to suggest this list should be kept in sync with rtti.c++.