Further refine capnp::SchemaLoader ownership
In the original work here I missed that the SchemaLoader is
needed during the construction of Worker::Script when using
the older service worker syntax. This change moves the ownership
of the SchemaLoader to the Worker::Script::Impl so that it is
available during construction, with the ContextGlobal having
a const reference to it.

The next step will be to move the capnp modules implementation
out from the internal project into workerd.