Add implementation of node::assert partialDeepStrictEqual
Another claude code special. This is a newer addition
to the node:assert module that does a deep strict equality
check but only for the properties present in the "expected"
object.

Just for fun, I had claude write both the implementation
and the tests. It did alright but I had to take it
through a few iterations to get it right. Once I
had claude actually analyze the existing implementation
in nodejs/node it finally managed to get it right.