Implement allowUnconfirmed in SQLite-backed implicit transactions
The idea here is that we’ll always lock the output gate if we have
writes, but release the output gate on the next turn of the event loop
if all of the writes to the transaction are `allowUnconfirmed` writes.
See comments in onWrite for details.

Claude wrote most of the tests.