Create enable_ctx_exports compat flag for ctx.exports.
Due to some weird edge cases, we can just enable ctx.exports for everyone in prod. But that means we need a compat flag for people to enable it. So this introduces that.

Since ctx.exports doesn't work in prod at all yet, we can safely *stop* enabling it for "experimental". This is the safe thing to do because we don't want people who are setting the "experimental" flag alone to accidentally rely on ctx.exports, when it won't work in prod unless they also set "enable_ctx_exports".