Ignore non-file URLs when building bundle-fs
Technically this would be a breaking change but it's still
early enough in the life of the vfs system that it's exceedingly
unlikely anyone is relying on this buggy behavior. If the module
name ends up being a non-file: URL we don't want to include it
in the file system bundle at all.