Add MemoryProtectionKeyScope
Utility class for capturing the current MPK for
the isolate (if enabled) and restoring it when
necessary. This is necessary for correct operation
when, for instance, we are writing to an ArrayBuffer
BackingStore from a coroutine outside the isolate
lock. Eventually we should see about making the
coroutine itself MPK-aware but this is a good
first step that keeps things relatively simple.