Address feedback: provide min read policy
Based on Harris' feedback, I've updated the implementation
of the ReadableStreamSourceKjAdapter to include a minimum
read policy. When the policy is IMMEDIATE, the adapter will
always return immediately after minimally fulfilling the
minimum read, even if more data is available and there is
more room in buffer we're reading into. If the policy is
OPPORTUNISTIC, then the adapter will attempt to fill the
buffer as much as possible (up until there is <= 512
bytes remaining in the buffer or we hit maxBytes) while
we are holding the isolate lock. The default policy is
OPPORTUNISTIC, with the idea being that we should do as
much work as possible while we have the isolate lock
before returning to the kj event loop.