Plumb allowUnconfirmed through SQL queries to the onWrite callback
It’s possible to mix writes that are confirmed with writes that are
unconfirmed.  We need to be able to distinguish which writes have what
unconfirmed status in the `onWrite` call.  The simplest way I could
figure out how to do this was to plumb whether a certain query is
confirmed or not all the way from where the queries are created
down into the SqliteDatabase layer, which can call back into the
ActorSqlite layer via the `onWrite` call.

Notably, this allows us to mix APIs that do support unconfirmed
writes (so far, just the async K/V API) with APIs that support
confirmed writes only (the `storage.sql` and `storage.kv` APIs) since
we default allowConfirmed to be false.