Disable `allowUnconfirmed` where we don’t want to support it (yet)
We want to enable `allowUnconfirmed` for puts and deletes in implicit
transactions.

This means disabling `allowUnconfirmed` for:
* anything in an explicit transaction.  The API for explicit
  transactions inherits the `allowUnconfirmed` option, so theoretically
  we could make all of the writes in a transaction be unconfirmed, and
  therefore the commit would be unconfirmed.  We may want to support
  this in the future
* `setAlarm`.  We may want to support this in the future.
* `deleteAll`.  This is probably difficult to support.

There is an edge case where multi-put creates its own explicit
transaction that will disable `allowUnconfirmed`.  This should
probably be fixed to use an implicit transaction instead.

When we override the `allowUnconfirmed` option to disable it, we log a
message.  I plan to look at these messages to see if they’re ever
actually logged.