Improve MemoryInputStream
Improve and expand the implementation of MemoryInputStream.
This will soon be used to replace the duplicative stream
implementations in: BlobInputStream, EmailPartInputStream,
BodyBufferInputStream, and a couple of test stream impls.

The improvements here include better handling of pumpTo,
and teeing, as well as allowing the stream to hold onto
its underlying backing.