containers: After the connection is done, we should throw an exception of the connection to the container service instead of sudden disconnections
This fixes a bug, when the body is set in a http request to a container,
if the TCP port call fails, there will be a Network Connection Lost,
instead of the original container connect error.

This does not happen on HEAD/GET requests, where a jsg error from the
container service is properly propagated.