Use ImplicitTxns by default for ActorSqlite multi-put
Normally, when you do a write with ActorSqlite and you’re not
otherwise in a transaction, you’ll make an ImplicitTxn that will
commit the next time the code goes async.  The multi-put code used an
ExplicitTxn, which would commit immediately.  It’s a subtle
difference, but we might as well get rid of it.

This change allows multi-puts to use `allowUnconfirmed`, though I plan
on using relaxing the change to enable `allowUnconfirmed` for
ExplicitTxns reasonably soon, so it’s not a huge win.