Move implementation of multi-put into SqliteKv
This allows us to use prepared statements for the SAVEPOINT and
RELEASE queries so we’re not parsing them all the time.  It comes at
the cost of using templates so that we don’t have to make copies of
`kj::Array<ActorSqlite::KeyValuePair>` values.  Probably worth it.

While I was here, I also updated the exception handling as [per
comments](https://github.com/cloudflare/workerd/pull/5253#discussion_r2411260094):
* `catch(…)` so we get all exceptions rather than only KJ ones
* catch and log any exceptions in the catch block so users get
  exceptions they can understand.