[o11y] Revert addition of getEventInfo to fix onset event reporting under RPC
This is incompatible with properly reporting custom event info under RPC, which
wasn't previously caught since it was only used for a subset of events. This
approach was previously introduced in
https://github.com/cloudflare/workerd/pull/5056, but is not required to report
the onset before JS execution starts – it is sufficient to do this before
delivered() is called, which is being done here.
By restoring the previous approach of adding event info in each
customEvent, we ensure that it is being set even under RPC, where
WorkerEntrypoint::customEvent doesn't run.