[o11y] Report TailStreamTarget stop signal when encountering exception
This should fix "The destination object for this tail session no longer exists"
errors when the tail stream client keeps sending events after the server has
encountered an exception. This scenario has not been reproduced, it is believed
to have happened in edge testing due to a race condition.
This requires extending the lifetime of the results object and the
CallContextHook that owns it so that it is accessible in both the promise where
we process events as well as its exception handler.