Implement allowUnconfirmed for explicit transactions
As far as I can tell from searching Kibana for the logs generated by
`disableAllowUnconfirmed`, allowUnconfirmed is never used.  However,
it was easy enough to implement this so I did.

The only thing left where we we disallow allowUnconfirmed is in
`deleteAll`, and I’m not sure how to not confirm the deleteAll, so I’m
going to leave it be for now.

Claude wrote the tests with a few human tweaks that were easier to do
manually.