Reset SQLite heap limit after test
The SQLite heap limits are process-wide, so setting it in a test will
affect all tests that run later.  I plan on adding more tests to this
file and while I _could_ ensure this test is always last in the file,
it’s better reset the value at the end of the test.

I had to take a direct dependency on sqlite3 because you can only
lower the hard heap limit with the `PRAGMA hard_heap_limit`.  To raise
the limit you must use the `sqlite3_hard_heap_limit64` function.