Remove backpressure debug warning
In cases where backpressure signals are being pathologically
ignored, the additional warning becomes too spammy to be useful.
Given that it's often out of the users control (as in the case
of a worker using next.js which suffers from this), the warning
becomes unhelpful additional noise. Just remove it.