Refactor SqliteKv’s multi-put to silence an errant clang-tidy lint
We have an internal clang-tidy link that checks for re-use of
exceptions.  Unfortunately, that internal check was confused by the
nested try/catch statements.

This refactor removes the try/catch statements and instead relies on
KJ primitives.  I’m not sure it’s easier to read, but it hews to the
KJ style of avoiding bare try/catch except where necessary.