Optimize the common header name handling
While the common header names and indices are defined in
the capnp header, they are unlikely to change often (if at
all). To avoid the overhead of building the hash map at
runtime and performing the hash lookups, we can hardcode
the common header names in an array and use direct indexing
to retrieve them. Should at least save a handful of CPU
cycles.