Remove jsg::ByteString
The type is largely obsolete and was being used only to
warn about invalid header names/values in Headers.
However, it incurred an additional overhead when
being created. This was especially non-sensical for
header names since we already validate those to be
proper HTTP token strings. Removing the type simplifies
code and helps boost overall performance.