Fix #include in the middle of the file.
This isn't supposepd to work. In particular, expanding the include _inside_ a namespace block puts all the symbols into the wrong namespace. However, the fact that this is a C header (wrapped in `extern "C" {`) coincidentally causes the misplaced symbols to end up linking correctly (because it disables complex C++ symbol names, effectively stripping the namespace prefixes). So it ended up working by accident, but we shouldn't rely on that.