[nfc] Require compatibility dates to be Tue-Thu
It's currently very easy to accidentally choose a default-on date for a
compatibility flag that lands on a weekend, or a Friday or a Monday,
none of which are ideal. We're most likely to have a full team on hand
to respond to any issues arising from default behavior changes on a work
day in the middle of the week, so this test codifies that.

Claude wrote this test, under my close supervision.