container: Add setEgressTcp capnp capability and setEgress
This is part of the workstream to give a container access to a WorkerEntrypoint.
How it works is that subrequest channels can implement a
writeServiceDesignator function that they can fill with metadata
appropiate to connect to the worker on the same machine. On workerd,
this should just be a name and an entrypoint.