Begin simplifying TypeWrapper
The jsg::TypeWrapper class has become a rather large and unwieldy
collection of base classes and mixins making it quite difficult to
compile, maintain, and even comprehend. When building, it can take
up to a full minute to (re)compile this single class due to the
complexity of its template metaprogramming. This change begins
to simplify by reducing the number of base classes and mixins,
moving parts of the functionality into the core TypeWrapper
class itself.