Use ReverseIoOwn in makeReentryCallback().
This ensures resources held by the callback can be freed more proactively when the IoContext shuts down, and also solves an unlikely scenario where the callback outlives the whole isolate (not just the IoContext), leading to a V8 fatal error.

This is difficult to reproduce. I have a load test I'm running that usually triggers this after a minute or two, which is obviously not appropriate for a unit test. But this change definitely fixes the problem.