[NFC] Update doxygen build and configuration files
I know that doxygen isn't everyone's favorite thing, but I've
found it useful, but it hasn't been updated in a bit. This
PR updates it and adds a just ds (doxygen-serve) command to
easily serve the generated docs for local browsing.

The doxygen build is added because doxygen needs to be
built with the correct clang version to avoid warnings and
errors.