[o11y] Do error checking in R2 earlier
Before #5790, this would have avoided "destructed WorkerTracer" warnings when we
would construct a WorkerInterface but not end up using it due to the errors
being detected later. However, this 1) does not apply for prod where R2 is
implemented differently and 2) no longer fixes a warning since the
WorkerInterface never has delivered() called. However, it is still good practice
to check for errors earlier and avoid memory allocations from the
getHttpClient() calls.
While this PR is not aiming for completion I also cleaned up a call in
web-socket.c++ that may be susceptible to the same issue.