Improve windows CI performance
Seems like we need to use Developer Mode to get real symlinks in windows.
From bazel profiles this seems to reduce symlink times by a ton.
A drive by here is to always upload bazel profiles and report disk usage
so that we have them available even for runs where a test flakes.